# TaskRonaut/backend/src/config/swaggerAll.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Taskronaut API",
    "version": "1.0.0",
    "description": "API documentation for Taskronaut"
  },
...
  "paths": {
    "/tasks": {
      "get": {
        "summary": "Holt alle Tasks",
        "description": "Gibt eine Liste aller vorhandenen Tasks aus der Datenbank zurück.",
        "tags": [
          "Task"
        ],
        "responses": {
          "200": {
            "description": "Eine Liste von Tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "taskID": {
                        "type": "integer"
                      },
                      "taskCreator": {
                        "type": "string"
                      },
                      "taskName": {
                        "type": "string"
                      },
                      "dueDate": {
                        "type": "string",
                        "format": "date"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Erstellt einen neuen Task",
        "description": "Fügt einen neuen Task in die Datenbank hinzu.",
        "tags": [
          "Task"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "taskCreator": {
                    "type": "string"
                  },
                  "taskCreatorID": {
                    "type": "integer"
                  },
                  "dueDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "responsible": {
                    "type": "string"
                  },
                  "taskDescription": {
                    "type": "string"
                  },
                  "taskName": {
                    "type": "string"
                  },
                  "priorities": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string"
                  },
                  "taskStatus": {
                    "type": "string"
                  },
                  "authorizedUser": {
                    "type": "string"
                  },
                  "comments": {
                    "type": "string"
                  },
                  "taskHistoryID": {
                    "type": "integer"
                  },
                  "boardID": {
                    "type": "integer"
                  },
                  "listID": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task erfolgreich erstellt"
          },
          "400": {
            "description": "Fehler beim Erstellen des Tasks"
          }
        }
      }
    },
    "/tasks/{id}": {
      "put": {
        "summary": "Updatet einen bestimmten Task",
        "description": "Aktualisiert die Eigenschaften eines bestehenden Tasks.",
        "tags": [
          "Task"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID des zu aktualisierenden Tasks"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "taskCreator": {
                    "type": "string"
                  },
                  "taskCreationDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "dueDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "responsible": {
                    "type": "string"
                  },
                  "taskDescription": {
                    "type": "string"
                  },
                  "taskName": {
                    "type": "string"
                  },
                  "priorities": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string"
                  },
                  "taskStatus": {
                    "type": "string"
                  },
                  "authorizedUser": {
                    "type": "string"
                  },
                  "comments": {
                    "type": "string"
                  },
                  "taskHistoryID": {
                    "type": "integer"
                  },
                  "boardID": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task erfolgreich aktualisiert"
          },
          "404": {
            "description": "Task nicht gefunden"
          }
        }
      },
      "delete": {
        "summary": "Löscht einen Task",
        "description": "Löscht einen Task basierend auf der übergebenen ID.",
        "tags": [
          "Task"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID des zu löschenden Tasks"
          }
        ],
        "responses": {
          "200": {
            "description": "Task erfolgreich gelöscht"
          },
          "404": {
            "description": "Task nicht gefunden"
          }
        }
      }
    },
    "/tasks/{id}/listId": {
      "patch": {
        "summary": "Aktualisiert die listId eines bestimmten Tasks",
        "description": "Setzt eine neue listId für den gegebenen Task basierend auf seiner ID.",
        "tags": [
          "Task"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID des zu aktualisierenden Tasks"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "listId": {
                    "type": "integer",
                    "description": "Neue List-ID für den Task"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task erfolgreich aktualisiert"
          },
          "400": {
            "description": "Fehlende listId im Request Body"
          },
          "404": {
            "description": "Task nicht gefunden"
          }
        }
      }
    },
    "/tasks/{user}": {
      "post": {
        "summary": "Get user-specific tasks",
        "description": "Retrieves tasks specific to a user based on their user ID and session.",
        "tags": [
          "Task"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the user."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "session": {
                    "type": "string",
                    "description": "The session token of the user."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully retrieved user-specific tasks.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "taskID": {
                        "type": "integer"
                      },
                      "taskCreator": {
                        "type": "string"
                      },
                      "taskName": {
                        "type": "string"
                      },
                      "dueDate": {
                        "type": "string",
                        "format": "date"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access due to invalid session."
          },
          "404": {
            "description": "User not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/getListIdFromTaskId/{id}": {
      "get": {
        "summary": "Get the list ID associated with a specific task ID.",
        "description": "This endpoint retrieves the list ID that corresponds to the given task ID.",
        "tags": [
          "Task"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The ID of the task.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the list ID.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "listId": {
                          "type": "string",
                          "example": "1234abcd"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Task ID not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Task ID not found."
                    },
                    "data": {
                      "type": "null"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal server error."
                    },
                    "data": {
                      "type": "null"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}